Move methods to extension files
