+++ piwigo/Album/AlbumSelector/SelectCategoryViewController.swift
@@ -31,15 +31,15 @@ protocol SelectCategoryImageRemovedDelegate: NSObjectProtocol {
     func didRemoveImage()
 }
 
-class SelectCategoryViewController: UIViewController, UITableViewDataSource, UITableViewDelegate {
+class SelectCategoryViewController: UIViewController {
 
     weak var delegate: SelectCategoryDelegate?
     weak var imageCopiedDelegate: SelectCategoryImageCopiedDelegate?
     weak var imageRemovedDelegate: SelectCategoryImageRemovedDelegate?
 
     var wantedAction: pwgCategorySelectAction = .none  // Action to perform after category selection
-    private var selectedCategoryId = Int32.min
-    private var updateOperations = [BlockOperation]()
+    var selectedCategoryId = Int32.min
+    var updateOperations = [BlockOperation]()
 
     // MARK: - Core Data Objects
     var user: User!
@@ -169,8 +169,7 @@ class SelectCategoryViewController: UIViewController, UITableViewDataSource, UIT
 
     @IBOutlet var categoriesTableView: UITableView!
     private var cancelBarButton: UIBarButtonItem?
-
-    private var albumsShowingSubAlbums = Set<Int32>()
+    var albumsShowingSubAlbums = Set<Int32>()
 
     
     // MARK: - Core Data Source
@@ -457,9 +456,9 @@ class SelectCategoryViewController: UIViewController, UITableViewDataSource, UIT
             dismiss(animated: true)
         }
     }
-    
 
-    // MARK: - UITableView - Header
+
+    // MARK: - TableView Main Header
     private func setTableViewMainHeader() {
         let headerView = SelectCategoryHeaderView(frame: .zero)
         switch wantedAction {
@@ -499,502 +498,12 @@ class SelectCategoryViewController: UIViewController, UITableViewDataSource, UIT
                                  text: NSLocalizedString("moveSeveralImages_selectAlbum", comment: "Please, select the album in which you wish to copy the photos."))
 
         default:
-            fatalError("Action not configured in setTableViewMainHeader().")
+            preconditionFailure("Action not configured in setTableViewMainHeader().")
         }
         categoriesTableView?.tableHeaderView = headerView
     }
 
-    private func getContentOfHeader(inSection section: Int) -> (String, String) {
-        var title = "", text = ""
-        switch wantedAction {
-        case .setAlbumThumbnail:
-            // 1st section —> Albums containing image
-            if section == 0 {
-                // Title
-                title = String(format: "%@\n", NSLocalizedString("tabBar_albums", comment:"Albums"))
-                text = inputImages.first?.albums?.count ?? 0 > 1 ?
-                    NSLocalizedString("categorySelection_one", comment:"Select one of the albums containing this image") :
-                    NSLocalizedString("categorySelection_current", comment:"Select the current album for this image")
-            } else {
-                // Text
-                text = NSLocalizedString("categorySelection_other", comment:"or select another album for this image")
-            }
-
-        default:
-            // 1st section —> Recent albums
-            if section == 0 {
-                // Do we have recent albums to show?
-                title = (recentAlbums.fetchedObjects ?? []).count > 0 ?
-                    NSLocalizedString("maxNberOfRecentAlbums>320px", comment: "Recent Albums") :
-                    NSLocalizedString("tabBar_albums", comment: "Albums")
-            } else {
-                // 2nd section
-                title = NSLocalizedString("categorySelection_allAlbums", comment: "All Albums")
-            }
-        }
-        return (title, text)
-    }
-    
-    func tableView(_ tableView: UITableView, heightForHeaderInSection section: Int) -> CGFloat {
-        let (title, text) = getContentOfHeader(inSection: section)
-        return TableViewUtilities.shared.heightOfHeader(withTitle: title, text: text,
-                                                        width: tableView.frame.size.width)
-    }
-
-    func tableView(_ tableView: UITableView, viewForHeaderInSection section: Int) -> UIView? {
-        let (title, text) = getContentOfHeader(inSection: section)
-        return TableViewUtilities.shared.viewOfHeader(withTitle: title, text: text)
-    }
-
-    
-    // MARK: - UITableView - Rows
-    func numberOfSections(in tableView: UITableView) -> Int {
-        switch wantedAction {
-        case .setAlbumThumbnail:
-            return 2
-        default:    // Present recent albums if any
-            let objects = recentAlbums.fetchedObjects ?? []
-            return 1 + (objects.isEmpty ? 0 : 1)
-        }
-    }
-
-    func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-        switch wantedAction {
-        case .setAlbumThumbnail:
-            if section == 0 {
-                return inputImages.first?.albums?.filter({$0.pwgID > 0}).count ?? 0
-            } else {
-                return (albums.fetchedObjects ?? []).count
-            }
-        default:    // Present recent albums if any
-            if ((recentAlbums.fetchedObjects ?? []).count > 0) && (section == 0) {
-                return (recentAlbums.fetchedObjects ?? []).count
-            } else {
-                return (albums.fetchedObjects ?? []).count
-            }
-        }
-    }
-    
-    func tableView(_ tableView: UITableView, heightForRowAt indexPath: IndexPath) -> CGFloat {
-        return 44.0;
-    }
-
-    func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-        
-        guard let cell = tableView.dequeueReusableCell(withIdentifier: "CategoryTableViewCell", for: indexPath) as? CategoryTableViewCell
-        else { preconditionFailure("Error: tableView.dequeueReusableCell does not return a CategoryTableViewCell!") }
-        var depth = 0
-        let albumData: Album
-        let hasRecentAlbums = (recentAlbums.fetchedObjects ?? []).isEmpty == false
-        switch indexPath.section {
-        case 0:
-            if wantedAction == .setAlbumThumbnail {
-                // Albums in which this image belongs to
-                var catId = inputAlbum.pwgID     // This album always exists in cache
-                if let catIds = inputImages.first?.albums?.compactMap({$0.pwgID}).filter({$0 > 0}),
-                   catIds.count > indexPath.row {
-                    catId = catIds[indexPath.row]
-                }
-                albumData = albumProvider.getAlbum(ofUser: user, withId: catId)!
-            } else if hasRecentAlbums {
-                // Recent albums
-                albumData = recentAlbums.object(at: indexPath)
-            } else {
-                // All albums
-                albumData = albums.object(at: indexPath)
-                if albumData.parentId > 0 {
-                    depth += albumData.upperIds.components(separatedBy: ",")
-                        .filter({ Int32($0) != albumData.pwgID }).count
-                }
-            }
-        default:
-            let albumIndexPath = IndexPath(item: indexPath.item, section: 0)
-            albumData = albums.object(at: albumIndexPath)
-            if albumData.parentId > 0 {
-                depth += albumData.upperIds.components(separatedBy: ",")
-                    .filter({ Int32($0) != albumData.pwgID }).count
-            }
-        }
-        
-        // No button if the user does not have upload rights
-        var buttonState: pwgCategoryCellButtonState = .none
-        let allAlbums: [Album] = albums.fetchedObjects ?? []
-        let filteredCat = allAlbums.filter({ user.hasAdminRights ||
-                                             userUploadRights.contains($0.pwgID) })
-        if filteredCat.count > 0 {
-            buttonState = albumsShowingSubAlbums.contains(albumData.pwgID) ? .hideSubAlbum : .showSubAlbum
-        }
-
-        // How should we present the category
-        cell.delegate = self
-        switch wantedAction {
-        case .setDefaultAlbum:
-            // The current default category is not selectable
-            if albumData.pwgID == inputAlbum.pwgID {
-                cell.configure(with: albumData, atDepth: depth, andButtonState: .none)
-                cell.albumLabel.textColor = .piwigoColorRightLabel()
-            } else {
-                // Don't present sub-albums in Recent Albums section
-                if hasRecentAlbums && (indexPath.section == 0) {
-                    cell.configure(with: albumData, atDepth: depth, andButtonState: .none)
-                } else {
-                    cell.configure(with: albumData, atDepth: depth, andButtonState: buttonState)
-                }
-            }
-        case .moveAlbum:
-            // User cannot move album to current parent album or in itself
-            if albumData.pwgID == 0 {  // Special case: upperCategories is nil for root
-                // Root album => No button
-                cell.configure(with: albumData, atDepth: depth, andButtonState: .none)
-                // Is the root album parent of the input album?
-                if inputAlbum.parentId == 0 {
-                    // Yes => Change text colour
-                    cell.albumLabel.textColor = .piwigoColorRightLabel()
-                }
-            }
-            else if hasRecentAlbums && (indexPath.section == 0) {
-                // Don't present sub-albums in Recent Albums section
-                cell.configure(with: albumData, atDepth: depth, andButtonState: .none)
-            }
-            else if albumData.pwgID == inputAlbum.parentId {
-                // This album is the parent of the input album => Change text colour
-                cell.configure(with: albumData, atDepth: depth, andButtonState: buttonState)
-                cell.albumLabel.textColor = .piwigoColorRightLabel()
-            }
-            else if albumData.upperIds.components(separatedBy: ",")
-                .compactMap({Int32($0)}).contains(inputAlbum.pwgID) {
-                // This album is a sub-album of the input album => No button
-                cell.configure(with: albumData, atDepth: depth, andButtonState: .none)
-                cell.albumLabel.textColor = .piwigoColorRightLabel()
-            } else {
-                // Not a parent of a sub-album of the input album
-                cell.configure(with: albumData, atDepth: depth, andButtonState: buttonState)
-            }
-        case .setAlbumThumbnail:
-            // Don't present sub-albums in first section
-            if indexPath.section == 0 {
-                cell.configure(with: albumData, atDepth: depth, andButtonState: .none)
-            } else {
-                cell.configure(with: albumData, atDepth: depth, andButtonState: buttonState)
-            }
-        case .setAutoUploadAlbum:
-            // Don't present sub-albums in Recent Albums section
-            if hasRecentAlbums && (indexPath.section == 0) {
-                cell.configure(with: albumData, atDepth: depth, andButtonState: .none)
-            } else {
-                cell.configure(with: albumData, atDepth: depth, andButtonState: buttonState)
-            }
-        case .copyImage, .copyImages, .moveImage, .moveImages:
-            // Don't present sub-albums in Recent Albums section
-            if hasRecentAlbums && (indexPath.section == 0) {
-                cell.configure(with: albumData, atDepth: depth, andButtonState: .none)
-            } else {
-                cell.configure(with: albumData, atDepth: depth, andButtonState: buttonState)
-            }
-            // Albums containing the image are not selectable
-            if commonCatIDs.contains(albumData.pwgID) {
-                cell.albumLabel.textColor = .piwigoColorRightLabel()
-            }
-
-        default:
-            break
-        }
-
-        cell.isAccessibilityElement = true
-        return cell
-    }
-
-    func tableView(_ tableView: UITableView, shouldHighlightRowAt indexPath: IndexPath) -> Bool {
-        // Retrieve album data
-        let albumData: Album
-        let hasRecentAlbums = (recentAlbums.fetchedObjects ?? []).count > 0
-        switch indexPath.section {
-        case 0:
-            // Provided album
-            if wantedAction == .setAlbumThumbnail {
-                var catId = inputAlbum.pwgID     // This album always exists in cache
-                if let catIds = inputImages.first?.albums?.compactMap({$0.pwgID}).filter({$0 > 0}),
-                   catIds.count > indexPath.row {
-                    catId = catIds[indexPath.row]
-                }
-                albumData = albumProvider.getAlbum(ofUser: user, withId: catId)!
-            } else if hasRecentAlbums {
-                // Recent albums
-                albumData = recentAlbums.object(at: indexPath)
-            } else {
-                // All albums
-                albumData = albums.object(at: indexPath)
-            }
-        default:
-            let albumIndexPath = IndexPath(item: indexPath.item, section: 0)
-            albumData = albums.object(at: albumIndexPath)
-        }
-
-        switch wantedAction {
-        case .setDefaultAlbum:
-            // The current default category is not selectable
-            debugPrint("••> albums: \(albumData.pwgID) and \(inputAlbum.pwgID)")
-            if albumData.pwgID == inputAlbum.pwgID {
-                return false
-            }
-            
-        case .moveAlbum:
-            // Do nothing if this is the input category
-            if albumData.pwgID == inputAlbum.pwgID {
-                return false
-            }
-            // User cannot move album to current parent album or in itself
-            if albumData.pwgID == 0 {  // upperCategories is nil for root
-                if inputAlbum.parentId == 0 {
-                    return false
-                }
-            } else if (albumData.pwgID == inputAlbum.parentId) ||
-                albumData.upperIds.components(separatedBy: ",")
-                .compactMap({Int32($0)}).contains(inputAlbum.pwgID) {
-                return false
-            }
-            
-        case .setAlbumThumbnail:
-            // The root album is not selectable (should not be presented but in case…)
-            if albumData.pwgID == 0 {
-                return false
-            }
-
-        case .setAutoUploadAlbum:
-            // The root album is not selectable (should not be presented but in case…)
-            if albumData.pwgID == 0 {
-                return false
-            }
-
-        case .copyImage, .copyImages, .moveImage, .moveImages:
-            // The root album is not selectable (should not be presented but in case…)
-            if albumData.pwgID == 0 {
-                return false
-            }
-            // Albums containing all the images are not selectable
-            if commonCatIDs.contains(albumData.pwgID) {
-                return false
-            }
-
-        default:
-            return false
-        }
-        return true;
-    }
-
-    
-    // MARK: - UITableViewDelegate Methods
-    func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
-        
-        tableView.deselectRow(at: indexPath, animated: true)
-
-        // Get selected category
-        let albumData: Album
-        let hasRecentAlbums = (recentAlbums.fetchedObjects ?? []).count > 0
-        switch indexPath.section {
-        case 0:
-            // Provided album
-            if wantedAction == .setAlbumThumbnail {
-                var catId = inputAlbum.pwgID     // This album always exists in cache
-                if let catIds = inputImages.first?.albums?.compactMap({$0.pwgID}).filter({$0 > 0}),
-                   catIds.count > indexPath.row {
-                    catId = catIds[indexPath.row]
-                }
-                albumData = albumProvider.getAlbum(ofUser: user, withId: catId)!
-            } else if hasRecentAlbums {
-                // Recent albums
-                albumData = recentAlbums.object(at: indexPath)
-            } else {
-                // All albums
-                albumData = albums.object(at: indexPath)
-            }
-        default:
-            let albumIndexPath = IndexPath(item: indexPath.item, section: 0)
-            albumData = albums.object(at: albumIndexPath)
-        }
-        
-        // Remember the choice
-        selectedCategoryId = albumData.pwgID
-
-        // What should we do with this selection?
-        switch wantedAction {
-        case .setDefaultAlbum:
-            // Do nothing if this is the current default category
-            if (albumData.pwgID == Int32.min) ||
-                (albumData.pwgID == inputAlbum.pwgID) { return }
-            
-            // Ask user to confirm
-            let title = NSLocalizedString("setDefaultCategory_title", comment: "Default Album")
-            let message:String
-            if albumData.pwgID == 0 {
-                message = String(format: NSLocalizedString("setDefaultCategory_message", comment: "Are you sure you want to set the album %@ as default album?"), NSLocalizedString("categorySelection_root", comment: "Root Album"))
-            } else {
-                message = String(format: NSLocalizedString("setDefaultCategory_message", comment: "Are you sure you want to set the album %@ as default album?"), albumData.name)
-            }
-            requestConfirmation(withTitle: title, message: message,
-                                forCategory: albumData, at: indexPath, handler: { [self] _ in
-                // Set new Default Album
-                self.delegate?.didSelectCategory(withId: albumData.pwgID)
-                // Return to Settings
-                self.navigationController?.popViewController(animated: true)
-            })
-
-        case .moveAlbum:
-            // Do nothing if this is the current default category
-            if albumData.pwgID == inputAlbum.pwgID { return }
-
-            // User must not move album to current parent album or in itself
-            if albumData.pwgID == 0 {  // upperCategories is nil for root
-                if inputAlbum.parentId == 0 {
-                    return
-                }
-            } else {
-                let parentIds = albumData.upperIds.components(separatedBy: ",")
-                if albumData.pwgID == inputAlbum.parentId ||
-                    parentIds.contains(where: { Int32($0) == inputAlbum.pwgID}) {
-                    return
-                }
-            }
-
-            // Ask user to confirm
-            let title = NSLocalizedString("moveCategory", comment: "Move Album")
-            let message = String(format: NSLocalizedString("moveCategory_message", comment: "Are you sure you want to move \"%@\" into the album \"%@\"?"), inputAlbum.name, albumData.name)
-            requestConfirmation(withTitle: title, message: message,
-                                forCategory: albumData, at: indexPath, handler: { [self] _ in
-                // Move album to selected category
-                self.moveCategory(intoCategory: albumData)
-            })
-
-        case .setAlbumThumbnail:
-            // Ask user to confirm
-            let title = NSLocalizedString("categoryImageSet_title", comment:"Album Thumbnail")
-            let message = String(format: NSLocalizedString("categoryImageSet_message", comment:"Are you sure you want to set this image for the album \"%@\"?"), albumData.name)
-            requestConfirmation(withTitle: title, message: message,
-                                forCategory: albumData, at: indexPath, handler: { [self] _ in
-                // Add category to list of recent albums
-                self.setRepresentative(for: albumData)
-            })
-
-        case .setAutoUploadAlbum:
-            // Do nothing if this is the root album
-            if albumData.pwgID == 0 { return }
-            
-            // Return the selected album ID
-            delegate?.didSelectCategory(withId: albumData.pwgID)
-            navigationController?.popViewController(animated: true)
-            
-        case .copyImage:
-            // Do nothing if this is the root album
-            if albumData.pwgID == 0 { return }
-            // Do nothing if the image already belongs to the selected album
-            if commonCatIDs.contains(albumData.pwgID) { return }
-
-            // Ask user to confirm
-            let title = NSLocalizedString("copyImage_title", comment:"Copy to Album")
-            let imageTitle = inputImages.first?.title.string ?? ""
-            let message = String(format: NSLocalizedString("copySingleImage_message", comment:"Are you sure you want to copy the photo \"%@\" to the album \"%@\"?"), imageTitle.isEmpty ? inputImages.first?.fileName ?? "-?-" : imageTitle, albumData.name)
-            requestConfirmation(withTitle: title, message: message,
-                                forCategory: albumData, at: indexPath, handler: { [self] _ in
-                // Display HUD
-                self.showHUD(withTitle: NSLocalizedString("copySingleImageHUD_copying", comment:"Copying Photo…"))
-                // Copy single image to selected album
-                self.copyImages(toAlbum: albumData)
-            })
-
-        case .moveImage:
-            // Do nothing if this is the root album
-            if albumData.pwgID == 0 { return }
-            // Do nothing if the image already belongs to the selected album
-            if commonCatIDs.contains(albumData.pwgID) { return }
-
-            // Ask user to confirm
-            let title = NSLocalizedString("moveImage_title", comment:"Move to Album")
-            let imageTitle = inputImages.first?.title.string ?? ""
-            let message = String(format: NSLocalizedString("moveSingleImage_message", comment:"Are you sure you want to move the photo \"%@\" to the album \"%@\"?"), imageTitle.isEmpty ? inputImages.first?.fileName ?? "-?-" : imageTitle, albumData.name)
-            requestConfirmation(withTitle: title, message: message,
-                                forCategory: albumData, at: indexPath) { [self] _ in
-                // Display HUD
-                self.showHUD(withTitle: NSLocalizedString("moveSingleImageHUD_moving", comment:"Moving Photo…"))
-                // Move single image to selected album
-                self.moveImages(toAlbum: albumData)
-            }
-
-        case .copyImages:
-            // Do nothing if this is the root album
-            if albumData.pwgID == 0 { return }
-            // Do nothing if the images already belong to the selected album
-            if commonCatIDs.contains(albumData.pwgID) { return }
-
-            // Ask user to confirm
-            let title = NSLocalizedString("copyImage_title", comment:"Copy to Album")
-            let message = String(format: NSLocalizedString("copySeveralImages_message", comment:"Are you sure you want to copy the photos to the album \"%@\"?"), albumData.name)
-            requestConfirmation(withTitle: title, message: message,
-                                forCategory: albumData, at: indexPath, handler: { [self] _ in
-                // Display HUD
-                self.showHUD(withTitle: NSLocalizedString("copySeveralImagesHUD_copying", comment: "Copying Photos…"),
-                             inMode: .determinate)
-                // Copy several images to selected album
-                DispatchQueue.global(qos: .userInitiated).async { [self] in
-                    self.copyImages(toAlbum: albumData)
-                }
-            })
-
-        case .moveImages:
-            // Do nothing if this is the root album
-            if albumData.pwgID == 0 { return }
-            // Do nothing if the images already belong to the selected album
-            if commonCatIDs.contains(albumData.pwgID) { return }
-
-            // Ask user to confirm
-            let title = NSLocalizedString("moveImage_title", comment:"Move to Album")
-            let message = String(format: NSLocalizedString("moveSeveralImages_message", comment:"Are you sure you want to move the photos to the album \"%@\"?"), albumData.name)
-            requestConfirmation(withTitle: title, message: message,
-                                forCategory: albumData, at: indexPath) { [self] _ in
-                // Display HUD
-                self.showHUD(withTitle: NSLocalizedString("moveSeveralImagesHUD_moving", comment: "Moving Photos…"),
-                             inMode: .determinate)
-                // Move several images to selected album
-                DispatchQueue.global(qos: .userInitiated).async { [self] in
-                    self.moveImages(toAlbum: albumData)
-                }
-            }
-
-        default:
-            break
-        }
-    }
-    
-    private func requestConfirmation(withTitle title:String, message:String,
-                                     forCategory albumData: Album, at indexPath:IndexPath,
-                                     handler:((UIAlertAction) -> Void)? = nil) -> Void {
-        let alert = UIAlertController(title: title, message: message, preferredStyle: .actionSheet)
-        let cancelAction = UIAlertAction(title: NSLocalizedString("alertCancelButton", comment: "Cancel"),
-                                         style: .cancel, handler: {_ in 
-                                            // Forget the choice
-                                            self.selectedCategoryId = Int32.min
-                                         })
-        let performAction = UIAlertAction(title: NSLocalizedString("alertYesButton", comment: "Yes"), style: .default, handler:handler)
-    
-        // Add actions
-        alert.addAction(cancelAction)
-        alert.addAction(performAction)
-
-        // Present popover view
-        alert.view.tintColor = .piwigoColorOrange()
-        if #available(iOS 13.0, *) {
-            alert.overrideUserInterfaceStyle = AppVars.shared.isDarkPaletteActive ? .dark : .light
-        } else {
-            // Fallback on earlier versions
-        }
-        alert.popoverPresentationController?.sourceView = categoriesTableView
-        alert.popoverPresentationController?.sourceRect = categoriesTableView.rectForRow(at: indexPath)
-        alert.popoverPresentationController?.permittedArrowDirections = [.left, .right]
-        present(alert, animated: true, completion: {
-            // Bugfix: iOS9 - Tint not fully Applied without Reapplying
-            alert.view.tintColor = .piwigoColorOrange()
-        })
-    }
-
+    // For presenting errors returned by Albums/Images actions
     func showError(_ error: Error?) {
         // Session logout required?
         if let pwgError = error as? PwgSessionError,
@@ -1043,88 +552,6 @@ class SelectCategoryViewController: UIViewController, UITableViewDataSource, UIT
             // Dismiss the view
             self.dismiss(animated: true, completion: {})
         }
-    }
-}
-
-
-// MARK: - NSFetchedResultsControllerDelegate
-extension SelectCategoryViewController: NSFetchedResultsControllerDelegate
-{
-    func controllerWillChangeContent(_ controller: NSFetchedResultsController<NSFetchRequestResult>) {
-        // Check that this update should be managed by this view controller
-        if (wantedAction == .setAlbumThumbnail) && (controller == recentAlbums) {
-            return
-        }
-        // Reset operation list
-        updateOperations = []
-        // Begin the update
-        categoriesTableView?.beginUpdates()
-    }
-    
-    func controller(_ controller: NSFetchedResultsController<NSFetchRequestResult>, didChange anObject: Any, at indexPath: IndexPath?, for type: NSFetchedResultsChangeType, newIndexPath: IndexPath?) {
-
-        // Check that this update should be managed by this view controller
-        if (wantedAction == .setAlbumThumbnail) && (controller == recentAlbums) {
-            return
-        }
-
-        // Initialisation
-        var hasAlbumsInSection1 = false
-        if controller == albums, categoriesTableView.numberOfSections == 2 {
-            hasAlbumsInSection1 = true
-        }
-
-        // Collect operation changes
-        switch type {
-        case .insert:
-            guard var newIndexPath = newIndexPath else { return }
-            if hasAlbumsInSection1 { newIndexPath.section = 1 }
-            updateOperations.append( BlockOperation { [weak self] in
-                debugPrint("••> Insert category item at \(newIndexPath)")
-                self?.categoriesTableView?.insertRows(at: [newIndexPath], with: .automatic)
-            })
-        case .update:
-            guard var indexPath = indexPath else { return }
-            if hasAlbumsInSection1 { indexPath.section = 1 }
-            updateOperations.append( BlockOperation {  [weak self] in
-                debugPrint("••> Update category item at \(indexPath)")
-                self?.categoriesTableView?.reloadRows(at: [indexPath], with: .automatic)
-            })
-        case .move:
-            guard var indexPath = indexPath,  var newIndexPath = newIndexPath else { return }
-            if hasAlbumsInSection1 {
-                indexPath.section = 1
-                newIndexPath.section = 1
-            }
-            updateOperations.append( BlockOperation { [weak self] in
-                debugPrint("••> Move category item from \(indexPath) to \(newIndexPath)")
-                self?.categoriesTableView?.moveRow(at: indexPath, to: newIndexPath)
-            })
-        case .delete:
-            guard var indexPath = indexPath else { return }
-            if hasAlbumsInSection1 { indexPath.section = 1 }
-            updateOperations.append( BlockOperation { [weak self] in
-                debugPrint("••> Delete category item at \(indexPath)")
-                self?.categoriesTableView?.deleteRows(at: [indexPath], with: .automatic)
-            })
-        @unknown default:
-            debugPrint("SelectCategoryViewController: unknown NSFetchedResultsChangeType")
-        }
-    }
-    
-    func controllerDidChangeContent(_ controller: NSFetchedResultsController<NSFetchRequestResult>) {
-        // Check that this update should be managed by this view controller
-        if (wantedAction == .setAlbumThumbnail) && (controller == recentAlbums) {
-            return
-        }
-
-        // Perform all updates
-        categoriesTableView?.performBatchUpdates { [weak self] in
-            self?.updateOperations.forEach { $0.start() }
-        }
-        
-        // End updates
-        categoriesTableView?.endUpdates()
     }
 }
