+++ piwigo/Upload/Upload Queue/UploadQueueViewController.swift
@@ -7,7 +7,6 @@
 //
 
 import CoreData
-import Photos
 import UIKit
 import piwigoKit
 import uploadKit
@@ -28,7 +27,7 @@ class UploadQueueViewController: UIViewController {
     // "private var diffableDataSource: DataSource!" replaced by below lines
     private var _diffableDataSource: NSObject? = nil
     @available(iOS 13.0, *)
-    private var diffableDataSource: DataSource {
+    var diffableDataSource: DataSource {
         if _diffableDataSource == nil {
             _diffableDataSource = configDataSource()
         }
@@ -418,306 +417,5 @@ class UploadQueueViewController: UIViewController {
     @objc func quitUpload() {
         // Leave Upload action and return to Albums and Images
         dismiss(animated: true)
-    }
-}
-
-
-// MARK: - UITableView - Diffable Data Source
-@available(iOS 13.0, *)
-extension UploadQueueViewController
-{
-    private func configDataSource() -> DataSource {
-        let dataSource = DataSource(tableView: queueTableView) { [self] (tableView, indexPath, objectID) -> UITableViewCell? in
-            // Get data source item
-            guard let upload = try? self.mainContext.existingObject(with: objectID) as? Upload else {
-                preconditionFailure("Managed item should be available")
-            }
-            // Configure cell
-            guard let cell = tableView.dequeueReusableCell(withIdentifier: "UploadImageTableViewCell", for: indexPath) as? UploadImageTableViewCell
-            else { preconditionFailure("Error: tableView.dequeueReusableCell does not return a UploadImageTableViewCell!") }
-            cell.configure(with: upload, availableWidth: Int(tableView.bounds.size.width))
-            return cell
-        }
-        dataSource.defaultRowAnimation = .fade
-        return dataSource
-    }
-}
-
-
-// MARK: - UITableViewDataSource Methods
-// Exclusively for iOS 12.x
-extension UploadQueueViewController: UITableViewDataSource
-{
-    func numberOfSections(in tableView: UITableView) -> Int {
-        if let sections = uploads.sections {
-            return sections.count
-        }
-        return 0
-    }
-
-    func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-        guard let sections = uploads.sections else {
-            fatalError("No sections in fetchedResultsController")
-        }
-        let sectionInfo = sections[section]
-        return sectionInfo.numberOfObjects
-    }
-    
-    func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-        guard let cell = tableView.dequeueReusableCell(withIdentifier: "UploadImageTableViewCell", for: indexPath) as? UploadImageTableViewCell
-        else { preconditionFailure("Error: tableView.dequeueReusableCell does not return a UploadImageTableViewCell!") }
-        cell.configure(with: uploads.object(at: indexPath),
-                       availableWidth: Int(tableView.bounds.size.width))
-        return cell
-    }
-}
-
-
-// MARK: - UITableViewDelegate Methods
-extension UploadQueueViewController: UITableViewDelegate
-{
-    // MARK: - UITableView - Headers
-    func tableView(_ tableView: UITableView, heightForHeaderInSection section: Int) -> CGFloat {
-        if #available(iOS 13.0, *) {
-            let sectionKey = SectionKeys(rawValue: diffableDataSource.snapshot()
-                                            .sectionIdentifiers[section]) ?? SectionKeys.Section4
-            return TableViewUtilities.shared.heightOfHeader(withTitle: sectionKey.name,
-                                                            width: tableView.frame.size.width)
-        } else {
-            // Fallback on earlier versions
-            var sectionName = SectionKeys.Section4.name
-            if let sectionInfo = uploads.sections?[section] {
-                let sectionKey = SectionKeys(rawValue: sectionInfo.name) ?? SectionKeys.Section4
-                sectionName = sectionKey.name
-            }
-            return TableViewUtilities.shared.heightOfHeader(withTitle: sectionName,
-                                                            width: tableView.frame.size.width)
-        }
-    }
-    
-    func tableView(_ tableView: UITableView, viewForHeaderInSection section: Int) -> UIView? {
-        guard let header = tableView.dequeueReusableHeaderFooterView(withIdentifier: "UploadImageHeaderView") as? UploadImageHeaderView
-        else { preconditionFailure("Error: tableView.dequeueReusableHeaderFooterView does not return a UploadImageHeaderView!") }
-        if #available(iOS 13.0, *) {
-            let sectionKey = SectionKeys(rawValue: diffableDataSource.snapshot().sectionIdentifiers[section]) ?? SectionKeys.Section4
-            header.config(with: sectionKey)
-            return header
-        } else {
-            // Fallback on earlier versions
-            if let sectionInfo = uploads.sections?[section] {
-                let sectionKey = SectionKeys(rawValue: sectionInfo.name) ?? SectionKeys.Section4
-                header.config(with: sectionKey)
-            } else {
-                header.config(with: SectionKeys.Section4)
-            }
-            return header
-        }
-    }
-    
-
-    // MARK: - UITableView - Rows
-    func tableView(_ tableView: UITableView, shouldHighlightRowAt indexPath: IndexPath) -> Bool {
-        return false
-    }
-    
-    func tableView(_ tableView: UITableView, trailingSwipeActionsConfigurationForRowAt indexPath: IndexPath) -> UISwipeActionsConfiguration? {
-        // Retreive upload object
-        guard let cell = tableView.cellForRow(at: indexPath) as? UploadImageTableViewCell,
-              let objectID = cell.objectID,
-              let upload = try? self.mainContext.existingObject(with: objectID) as? Upload
-        else { preconditionFailure("Managed object should be available") }
-        
-        // Create retry upload action
-        let retry = UIContextualAction(style: .normal, title: nil,
-                                       handler: { action, view, completionHandler in
-            UploadManager.shared.backgroundQueue.async {
-                UploadManager.shared.resumeFailedUpload(withID: upload.localIdentifier)
-                UploadManager.shared.findNextImageToUpload()
-            }
-            completionHandler(true)
-        })
-        retry.backgroundColor = .piwigoColorBrown()
-        retry.image = UIImage(named: "swipeRetry.png")
-        
-        // Create trash/cancel upload action
-        let cancel = UIContextualAction(style: .normal, title: nil,
-                                        handler: { action, view, completionHandler in
-            let savingContext = upload.managedObjectContext
-            savingContext?.delete(upload)
-            try? savingContext?.save()
-            UploadManager.shared.backgroundQueue.async {
-                UploadManager.shared.resumeFailedUpload(withID: upload.localIdentifier)
-                UploadManager.shared.findNextImageToUpload()
-            }
-            completionHandler(true)
-        })
-        cancel.backgroundColor = .red
-        cancel.image = UIImage(named: "swipeCancel.png")
-
-        // Associate actions
-        switch upload.state {
-        case .preparing, .prepared, .uploading, .uploaded, .finishing:
-            return UISwipeActionsConfiguration(actions: [retry])
-        case .preparingError, .uploadingError, .finishingError:
-            return UISwipeActionsConfiguration(actions: [retry, cancel])
-        case .waiting, .preparingFail, .formatError, .uploadingFail, .finishingFail, .finished, .moderated:
-            return UISwipeActionsConfiguration(actions: [cancel])
-        }
-    }
-
-    @objc func applyUploadProgress(_ notification: Notification) {
-        if let localIdentifier =  notification.userInfo?["localIdentifier"] as? String, !localIdentifier.isEmpty ,
-           let progressFraction = notification.userInfo?["progressFraction"] as? Float,
-           let visibleCells = queueTableView.visibleCells as? [UploadImageTableViewCell],
-           let cell = visibleCells.first(where: {$0.localIdentifier == localIdentifier}) {
-            debugPrint("••> progressFraction = \(progressFraction) in applyUploadProgress()")
-            cell.uploadingProgress?.setProgress(progressFraction, animated: true)
-        }
-    }
-}
-
-
-// MARK: - Uploads Provider NSFetchedResultsControllerDelegate
-extension UploadQueueViewController: NSFetchedResultsControllerDelegate
-{
-    @available(iOS 13.0, *)
-    func controller(_ controller: NSFetchedResultsController<NSFetchRequestResult>,
-                    didChangeContentWith snapshot: NSDiffableDataSourceSnapshotReference) {
-        // Expected controller?
-        guard controller == uploads else { return }
-        
-        // Data source configured?
-        guard let dataSource = queueTableView.dataSource as? UITableViewDiffableDataSource<String, NSManagedObjectID>
-        else { preconditionFailure("The data source has not implemented snapshot support while it should") }
-        
-        // Loop over all items
-        var snapshot = snapshot as NSDiffableDataSourceSnapshot<String, NSManagedObjectID>
-        let currentSnapshot = dataSource.snapshot() as NSDiffableDataSourceSnapshot<String, NSManagedObjectID>
-        var reloadIdentifiers: [NSManagedObjectID] = snapshot.itemIdentifiers
-        snapshot.itemIdentifiers.forEach({ itemIdentifier in
-            // Will this item keep the same indexPath?
-            guard let currentRow = currentSnapshot.indexOfItem(itemIdentifier),
-                  let row = snapshot.indexOfItem(itemIdentifier),
-                  row == currentRow,
-                  let currentSectionIdentifier = currentSnapshot.sectionIdentifier(containingItem: itemIdentifier),
-                  let currentSection = currentSnapshot.indexOfSection(currentSectionIdentifier),
-                  let sectionIdentifier = snapshot.sectionIdentifier(containingItem: itemIdentifier),
-                  let section = snapshot.indexOfSection(sectionIdentifier),
-                  section == currentSection
-            else { return }
-            reloadIdentifiers.removeAll(where: {$0 == itemIdentifier})
-            // Update upload state
-            let indexPath = IndexPath(row: row, section: section)
-            if let upload = try? controller.managedObjectContext.existingObject(with: itemIdentifier) as? Upload,
-               let cell = queueTableView.cellForRow(at: indexPath) as? UploadImageTableViewCell {
-                // Only update label
-                cell.uploadInfoLabel.text = upload.stateLabel
-            }
-        })
-
-        // Any item to reload/reconfigure?
-        if reloadIdentifiers.isEmpty == false {
-            // Animate only a non-empty UI
-            let shouldAnimate = queueTableView.numberOfSections != 0
-            if #available(iOS 15.0, *) {
-                snapshot.reconfigureItems(Array(reloadIdentifiers))
-            } else {
-                snapshot.reloadItems(Array(reloadIdentifiers))
-            }
-            dataSource.apply(snapshot as NSDiffableDataSourceSnapshot<String, NSManagedObjectID>,
-                             animatingDifferences: shouldAnimate)
-        }
-        
-        // Update the navigation bar
-        self.updateNavBar()
-        
-        // If all upload requests are done, delete all temporary files (in case some would not be deleted)
-        if snapshot.numberOfItems == 0 {
-            // Delete remaining files from Upload directory (if any)
-            UploadManager.shared.backgroundQueue.async {
-                UploadManager.shared.deleteFilesInUploadsDirectory()
-            }
-            // Close the view when there is no more upload request to display
-            self.dismiss(animated: true, completion: nil)
-        }
-    }
-    
-    // Exclusively for iOS 12.x
-    func controllerWillChangeContent(_ controller: NSFetchedResultsController<NSFetchRequestResult>) {
-        queueTableView.beginUpdates()
-    }
-    
-    // Exclusively for iOS 12.x
-    func controller(_ controller: NSFetchedResultsController<NSFetchRequestResult>, didChange sectionInfo: NSFetchedResultsSectionInfo, atSectionIndex sectionIndex: Int, for type: NSFetchedResultsChangeType) {
-        debugPrint("    > sectionInfo:", sectionInfo)
-
-        switch type {
-        case .insert:
-            debugPrint("insert section… at", sectionIndex)
-            queueTableView.insertSections(IndexSet(integer: sectionIndex), with: .automatic)
-        case .delete:
-            debugPrint("delete section… at", sectionIndex)
-            queueTableView.deleteSections(IndexSet(integer: sectionIndex), with: .automatic)
-        case .move, .update:
-            fallthrough
-        @unknown default:
-                fatalError("UploadQueueViewControllerOld: unknown NSFetchedResultsChangeType")
-        }
-    }
-
-    // Exclusively for iOS 12.x
-    func controller(_ controller: NSFetchedResultsController<NSFetchRequestResult>, didChange anObject: Any, at indexPath: IndexPath?, for type: NSFetchedResultsChangeType, newIndexPath: IndexPath?) {
-        
-        switch type {
-        case .insert:
-            guard let newIndexPath = newIndexPath else { return }
-            debugPrint("insert… at", newIndexPath)
-            queueTableView.insertRows(at: [newIndexPath], with: .automatic)
-        case .delete:
-            guard let oldIndexPath = indexPath else { return }
-            debugPrint("delete… at", oldIndexPath)
-            queueTableView.deleteRows(at: [oldIndexPath], with: .automatic)
-        case .move:
-            guard let oldIndexPath = indexPath else { return }
-            guard let newIndexPath = newIndexPath else { return }
-            debugPrint("move… from", oldIndexPath, "to", newIndexPath)
-            queueTableView.deleteRows(at: [oldIndexPath], with: .fade)
-            queueTableView.insertRows(at: [newIndexPath], with: .fade)
-        case .update:
-            guard let oldIndexPath = indexPath, let upload = anObject as? Upload,
-                  let cell = queueTableView.cellForRow(at: oldIndexPath) as? UploadImageTableViewCell
-            else { return }
-            debugPrint("update… at", oldIndexPath)
-            if (newIndexPath == nil) || (newIndexPath == oldIndexPath) {        // Regular update
-                cell.uploadInfoLabel.text = upload.stateLabel
-                if [.preparingError, .preparingFail, .formatError,
-                    .uploadingError, .uploadingFail, .finishingError].contains(upload.state) {
-                    // Display error message
-                    cell.imageInfoLabel.text = cell.errorDescription(for: upload)
-                }
-            } else {
-                queueTableView.deleteRows(at: [oldIndexPath], with: .automatic)
-                queueTableView.insertRows(at: [newIndexPath!], with: .automatic)
-            }
-        @unknown default:
-            fatalError("UploadQueueViewControllerOld: unknown NSFetchedResultsChangeType")
-        }
-    }
-    
-    // Exclusively for iOS 12.x
-    func controllerDidChangeContent(_ controller: NSFetchedResultsController<NSFetchRequestResult>) {
-        // Perform tableView updates
-        queueTableView.endUpdates()
-        queueTableView.layoutIfNeeded()
-
-        // If all upload requests are done, delete all temporary files (in case some would not be deleted)
-        if (uploads.fetchedObjects ?? []).count == 0 {
-            // Delete remaining files from Upload directory (if any)
-            UploadManager.shared.deleteFilesInUploadsDirectory()
-            // Close the view when there is no more upload request to display
-            self.dismiss(animated: true, completion: nil)
-        } else {
-            updateNavBar()
-        }
     }
 }
